Stickler fix 1012
