+++ src/index.js
@@ -1,35 +1,11 @@
 import React from 'react';
-import { createStore } from 'redux';
 import ReactDOM from 'react-dom';
 import './index.css';
 import { Provider } from 'react-redux';
 import App from './components/App';
 import * as serviceWorker from './serviceWorker';
-import rootReducer from './reducers/books';
+import store from '../src/store';
 
-let store;
-const initState = [
-  {
-    key: new Date().getTime() + 1,
-    title: 'Woof stories',
-    author: 'Dulce',
-    category: 'Action',
-  },
-  {
-    key: new Date().getTime() + 2,
-    title: 'Kitchen woofs',
-    author: 'Dulce',
-    category: 'Learning',
-  },
-  {
-    key: new Date().getTime() + 3,
-    title: 'Be the Woof',
-    author: 'Dulce',
-    category: 'Biography',
-  }
-];
-
-export default store = createStore(rootReducer, initState);
 
 ReactDOM.render(
   <Provider store={store}>
